Add some initial structure to the repo
The project is at its initial stages and we have been getting a brilliant flurry of PR's contributing various guides, standards and documents.

However they are all adding artifacts that just go into the root or the project. That's fine whilst we only have a handful, but as the project grows this is likely to become a burden. If nothing else we want the project's root README to be front and centre when someone navigates to https://github.com/DEFRA/software-development-standards, but the more files that exist at the root, the further down the page it'll be pushed.

I also think having some initial structure will help those looking to contribute because it will give them an indication of where their idea best fits in the project.

So this change adds some initial folders, each with their own README to indicate their purpose/content.

If this is accepted I'm more than happy to rebase and refactor all existing PR's to fit the structure, so please don't let be a concern 😄

---

WIP