Add developer workflows guide [coding practise]
This provides a guide on developer workflows which comes under the tag of 'coding practise'.

We are setting Git as our standard for version control, and that developers should release their code as early as possible to Github.

As such it is hoped that those new to Git will find this guide useful, and that it will support them in adopting Git.