Add more detail about the PR process inc. approval (#11)
There are some additional features in GitHub's PR functionality that we would benefit from taking advantage of

**Assigning a person**

Ideally all PR's should be assigned to a person, and that person should be who contributed or at least is leading on the contribution. This makes it easier for those reviewing to see who to direct comments and feedback to.

**Requesting reviews**

Requesting a review is a clear statement that your PR is ready to be reviewed. It's also a great way of letting those involved in the group know if you request DEFRA/sds-group as GitHub will automatically send a request out to those in the group.

**Using labels**

You can assign labels to a PR, and these can quickly highlight the nature of it even before reviewers have read through the intro.

For example this change has been flagged as *housekeeping* because its not actually related to new content such as a new standard, but is more about some improvements to the project itself.

---

This change adds instructions on using these to our CONTRIBUTING.md, along with more details about the process including how final approval works.