Move document to standards plus tweaks
With the introduction of a structure to the project, this change takes advantage and moves the document to the 'standards' folder.

Also I have used my knowledge of markdown to make some tweaks, namely using full code blocks with the `csharp` annotation to improve how the code is displayed, and embedding links into the text where it looked applicable.