Add general guidance on writing code
In PR #12 @Qbtron22 created some documents to cover security guidance, principles and standards. The feedback from @DEFRA/sds-group was that this was great, just they felt the guidance had a wider scope than just security.

So this PR takes the guidance document and adds it as general guidance on writing code. In fact because it is so applicable to most things it embeds it in the guidance section's main README.

🏆 All credit to @Qbtron22 for pulling together the original guidance 🏆