Add standard for READMEs
This PR adds a standard for READMEs in project repos.

Following on from our group webex on 25 November, it is based off of the future farming docs (https://github.com/DEFRA/ffc-development-guide/blob/master/docs/documentation-standards.md) but with some modifications.

The intent is that we will iterate to include more documentation about documentation, but this is a starting point.