Add process on what to do when creds. are exposed
Last week we on the [ruby services team](https://github.com/DEFRA/ruby-services-team) inadvertently pushed some AWS credentials to GitHub. This set off a chain of events, which were all resolved in the end, but which highlighted we don't have a clear process or guidance for what to do.

So this change will add a first version of the process. It may be AWS and GitHub centric as those were the main things involved in this experience. But it at least gives us something to build on in the future.