Copy across versioning suggestion
Include in PR #8 was the recommendation that all standards should be versioned, and where they are used the version should be referred to.

Again this would be some that applicable to all standards, not just PL/SQL so it has been migrated here to allow PR #8 to progress, and this suggestion be discussed in isolation.

For reference, as part of the copy across I have made the following changes

- slight grammatical tweaks to fit the new context
- markdown formatting tweaks, namely using bold to highlight the examples
- removing the line breaks within sentances. There is no need to break lines in markdown at 80 chars, particularly when it is solely for display in GitHub