Remove derivation section
To be honest, I'm not sure of where we got to regards discussion in the PR on this section.

It was certainly highlighted by @bensagar-ea that it could be expanded to include Defra and its other parts, plus external standards and guidance.

This could be ready as a generally applicable to all standards, but then the specifics of each standard might be very different.

One might be 100% a external standard, another might be based on one with our own changes, or 100% of our own making.

So knowing that this section if it was to be retained would need more work, but concious the goal is to get this PR done and dusted I've just gone with deleting it for now.

I've based this on the convention the other standards have followed, which is just to talk about the conventions and rules without any pre-amble. Secondly the project README also contains some general statements similar to this, plus we go into more detail in our [principles](https://github.com/DEFRA/software-development-standards/blob/master/principles/README.md).