Add content from original PL/SQL PR
This is the content from the original PR for PL/SQL standards. I've had to do a certain amount of editing to the format and content of the pre-amble as part of this.

The original PR also brought in the idea of all standards needing to be versioned, and these versions having to be referred to, which is something we've also extracted out as a new PR (see https://github.com/DEFRA/software-development-standards/pull/25).

As that has yet to be decided I've omitted that content. Also the original contribution marked the content up as headings. That's where I have re-formatted things so we have actual headings, and the extracted text then forms the section content.