Delete appendix covering auto-format in Toad
As per the discussion in the PR, this was felt to be more appropiate as a guide rather than being in the standard.

So the guidance can be discussed in isolation, separate from these standards I have created a new PR (https://github.com/DEFRA/software-development-standards/pull/26) and extracted the content to there.

So this change covers deleting them from the standard.