Add guidance for auto formatting in TOAD (#26)
This stems from an agreed refactoring of PR #8 (PL/SQL coding standards). It was highlighted in that PR some of the content would be more appropriate as guidance, rather than being a standard we must all follow.

So it was suggested to extract and move it into the `/guidance` section of the project.

SO the content of this PR comes from @nigejohnson, structure and how the changes were made, was me @Cruikshanks

I had to do a certain amount of editing to the format and content of the pre-amble as part of this.

The original PR also brought in the idea of all standards needing to be versioned, and these versions having to be referred to, which is something we've also extracted out as a new PR (see https://github.com/DEFRA/software-development-standards/pull/25).

As that has yet to be decided I've omitted that content. Also the original contribution marked the content up as headings. That's where I have re-formatted things so we have actual headings, and the extracted text then forms the section content.