Add back in standards derivation text
From the feedback in the PR it looks like I got the wrong end of the stick when it came to the derivation section. It was fine to stay in.

So this change brings it back. The one tweak (apart from removiung line breaks) I have made is to drop the section title and just have it as the opening paragraph to the standard.