Add guidance for auto formatting in Eclipse
This stems from an agreed refactoring of PR #5 (Java coding standards). It was highlighted in that PR some of the content would be more appropriate as guidance, rather than being a standard we must all follow.

It was suggested to extract and move it into the `/guidance` section of the project.

So the content of this PR comes from @nigejohnson and questions and comments should be directed to him 😁

Anything regards structure and how the changes were made, feel free to fire at me @Cruikshanks 😰