Add content from original Java PR
This is the content from the original PR for Java standards. I've had to do a certain amount of editing to the format and content of the pre-amble as part of this.

The original PR also brought in the idea of all standards needing to be versioned, and these versions having to be referred to, which is something we've also extracted out as a new PR (see #25).

As that has yet to be decided I've omitted that content.