RESTful API standards.
Just copied over (with some minor formatting changes) from what I originally wrote in 2015, so now superseded in some areas by the government API guidance.

The purpose of putting these in here now is to get a review of what needs changing, or whether the whole thing can be dropped.