Using unmanaged devices.
Some guidance on unmanaged devices that I started drafting a long while ago but never completed as someone else got given the job of doing it, but I never saw anything from that so thought we could start reviving this one.

It is out of date but probably could still do with some initial discussion/review.