Add guidance for auto formatting in Eclipse (#29)
This stems from an agreed refactoring of PR #5 (Java coding standards). It was highlighted in that PR some of the content would be more appropriate as guidance, rather than being a standard we must all follow.

It was suggested to extract and move it into the `/guidance` section of the project.

So the content comes from the original PR for Java standards. I've had to do a certain amount of editing to the format and content of the pre-amble as part of this.