Add example effective dates to a standard
Some of the PR's have include discussions about ensuring teams are not penalised because they are having to work with code bases that do not follow whatever is the current latest standard as defined here.

We have discussed manually versioning our standards. Concensus however was as we are using source control, manually applying a version seems an unnecessary burden. Another solution has been to use dates. This solution also fits in with one of our current principles for standards

> ...all standards must define the period of time that they are effective and how they have changed over time

The problem is we don't as yet have an actual example of how to go about this. Hence this proposed change to the Java coding standard, to demonstrate (and collaborate) on an approach that can then be retroactively applied to our existing standards, and used for any new ones in the future.