First draft of style and writing guide for standards
As discussed in the meeting today. This should cover the points we talked about, including Markdown and the concept of a standard being "easily testable".

This covers some basic formatting guidance, and some general writing advice for the tone and style of the guides. The goal is to make our standards easy to understand, and therefore easy to adopt.

I've also linked to some more extensive guidelines on writing for the web. These are written for a slightly different audience but I think the principles still apply. (Also, I wrote some of them back when I used to be a content person.)