Guide for new starters in development (#31)
This was raised in https://github.com/DEFRA/software-development-standards/issues/30 - luckily I was asked to get one of these together a while ago, and have been working on it in the background! But this repo does seem like a natural home for it.

The guide is mostly about getting access to various tools, sharing links to various standards, and encouraging people to get on Slack and come to meetups.